lunches + cutleries

+++ app/src/main/java/com/example/kursworkapplication/data/Cutlery.java
@@ -5,7 +5,7 @@ public class Cutlery {
     private int count;
     private String name;
     private String userLogin;
-    private Long order_id;
+    private int order_id;
 
     @Override
     public String toString(){
@@ -45,11 +45,11 @@ public class Cutlery {
         this.userLogin = userLogin;
     }
 
-    public Long getOrder_id() {
+    public int getOrder_id() {
         return order_id;
     }
 
-    public void setOrder_id(Long order_id) {
+    public void setOrder_id(int order_id) {
         this.order_id = order_id;
     }
 }
